+++ CHANGELOG.md
@@ -38,6 +38,8 @@ All notable changes to this project will be documented in this file.
 
 - Converted the validate_merge_coverage_data job script from PyDeequ to Pointblank.
 
+- Converted ascwds_filled_posts_calculator utils folder to polars within Clean Ind CQC Job.
+
 ### Changed
 - Remove interim/demo model preprocessing/retraining code.